requirements.txt
google-genai>=0.8.0
httpx>=0.27.0
python-dotenv>=1.0.0
fal-client>=0.5.0
pytest>=7.0.0
